pyproject.toml
[tool.poetry]
name = "crossfire"
version = "0.1.0"
description = ""
authors = ["Felipe Barros <felipe.b4rros@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.5"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
